skin_analyzer/requirements.txt
tensorflow>=2.15.0
keras>=2.15.0
opencv-python>=4.9.0.80
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
joblib>=1.3.0
